CAT-13 type class-5b/5c records as Class5b or Class5c
CatiaConsolidatedClass5b5cRecord.class is Class5b (0x5b) or Class5c
(0x5c). CADIR JSON still emits the class byte. The post-hoc 0x5b|0x5c
check is deleted.

Checks: cargo fmt --all && cargo check -q --workspace --tests
Hook-Bypass-Reason: the mission requires --no-verify for each non-final item; the full gate runs once at the end.

+++ crates/cadmpeg-codec-catia/src/families/b2/tests.rs
@@ -608,7 +608,7 @@ fn b2_class5b5c_parser_retains_complete_source_local_control_lanes() {
     assert_eq!(
         records
             .iter()
-            .map(|record| record.class)
+            .map(|record| u8::from(record.class))
             .collect::<Vec<_>>(),
         [0x5b, 0x5c, 0x5b]
     );
@@ -637,7 +637,7 @@ fn b2_class5b5c_parser_retains_complete_source_local_control_lanes() {
     assert_eq!(
         crate::families::b2::records::b2_class5b5c_records(&invalid_flag)
             .iter()
-            .map(|record| record.class)
+            .map(|record| u8::from(record.class))
             .collect::<Vec<_>>(),
         [0x5c, 0x5b]
     );
